requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
bottle==0.13.4
build==1.3.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
chromadb==1.1.0
click==8.3.0
coloredlogs==15.0.1
comm==0.2.3
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
durationpy==0.10
executing==2.2.1
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
gitpython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
hf-transfer==0.1.9
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.1
humanfriendly==10.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.6.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.8.1
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.11
langsmith==0.4.31
llama-cpp-python==0.3.16
macholib==1.16.3
markdown-it-py==4.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
melodycomp @ file:///Users/danvicente/Desktop/melodycomp
mido==1.2.6
mlx==0.29.2
mlx-lm==0.28.2
mlx-metal==0.29.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy-extensions==1.1.0
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.3
oauthlib==3.3.1
onnxruntime==1.22.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
polars==1.34.0
polars-runtime-32==1.34.0
posthog==5.4.0
pretty-midi==0.2.11
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
proxy-tools==0.1.0
psutil==7.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.11.0
pydeck==0.9.1
pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyobjc-core==11.1
pyobjc-framework-cocoa==11.1
pyobjc-framework-quartz==11.1
pyobjc-framework-security==11.1
pyobjc-framework-webkit==11.1
pyparsing==3.2.5
pypika==0.48.9
pyproject-hooks==1.2.0
pysmartdl==1.3.4
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywebview==5.4
pyyaml==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
ruff==0.13.3
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.43
stack-data==0.6.3
streamlit==1.50.0
streamlit-desktop-app==0.3.4
sympy==1.14.0
symusic==0.5.8
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
typer==0.19.2
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.14
websocket-client==1.8.0
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
